Notifications API moved from core to notifications app. Test cases updates and better strcturing.
