+++ core/tests/__init__.py
@@ -1,0 +1,1 @@
+# Required for pytest to discover tests in subdirectories
