+++ core/tests/test_utils.py
@@ -1,0 +1,20 @@
+from random import randint
+
+from core.utils import generate_id, number_to_random_token, token_to_number
+
+
+def test_id_token_conversion():
+    for _ in range(100):
+        num = randint(0, 10000)  # nosec: B311
+        tok = number_to_random_token(num)
+        assert isinstance(tok, str)
+        assert len(tok) > 6
+        assert num == token_to_number(tok)
+
+def test_generate_id():
+    id1 = generate_id()
+    assert len(id1) == 12
+    assert id1[0].isalpha()
+    assert id1.isalnum()
+    ids = {generate_id() for _ in range(1000)}
+    assert len(ids) == 1000
