+++ pyproject.toml
@@ -118,7 +118,7 @@ fail_under = 80
 
 
 [tool.bandit]
-exclude_dirs = ["./.venv", "tests.py", "sbom_format_schemas"]
+exclude_dirs = ["./.venv", "tests.py", "sbom_format_schemas", "tests", "node_modules"]
 # B101 = ignore assert_used
 # B404 = allow use of sub process module
 # B603 = allow calls to Popen with shell=False
